validators: run no-scrolling check after content type validator
