+++ lib/plugins/validators/async/21_checkContentType.js
@@ -2,6 +2,7 @@ const utils = require('../../../utils');
 const urlLib = require('url');
 const multimedia = require('../html5_multimedia');
 const mediaPlugin = require('../media');
+const player_no_scrolling = require('../player_no_scrolling');
 
 module.exports = {
 
@@ -27,7 +28,8 @@ module.exports = {
             }
 
             multimedia.prepareLink(link, options);
-            mediaPlugin.prepareLink(link, options);            
+            mediaPlugin.prepareLink(link, options);
+            player_no_scrolling.prepareLink(link, options);
 
             cb();
         }
