+++ lib/plugins/validators/sync/12_player_no_scrolling.js
@@ -1,13 +1,8 @@
+const player_no_scrolling = require('../player_no_scrolling');
+
 module.exports = {
 
-    prepareLink: function(link) {
-
-        if (link.href 
-            && (link.type === CONFIG.T.text_html || link.accept === CONFIG.T.text_html 
-                || (Array.isArray(link.accept) && link.accept.length == 1 && link.accept.indexOf(CONFIG.T.text_html) >-1)) 
-            && link.rel.indexOf('player') > -1 && link.media && link.media['aspect-ratio'] && !link.media.scrolling) {
-
-            link.media.scrolling = 'no';
-        }
+    prepareLink: function(link, options) {
+        player_no_scrolling.prepareLink(link, options);
     }
 };
